E1_GestionDonnees/api/requirements.txt
# API REST
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Base de données
sqlalchemy==2.0.25

# Sécurité
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilitaires
python-dotenv==1.0.1 